pyproject.toml
[project]
name = "screen-supervisor"
version = "0.1.0"
description = "全天候定时截屏并按日期归档的监控工具"
readme = "README.md"
authors = [{name = "Screen Supervisor"}]
requires-python = ">=3.11"
dependencies = [
    "mss>=9.0.1",
    "pillow>=10.3.0",
    "typer>=0.12.3",
    "pydantic>=2.7.1",
    "rich>=13.7.0"
]

[project.scripts]
screen-supervisor = "screen_supervisor.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.4.0"
]
